Initial code drop
First drop of code for a tool for automatically generating REST
documentation. A combination of Javadoc comments and metadata provided
by Spring's annotations are used to generate the documentation.

The primary focus of the tool is to avoid the need to repeat
information in multiple places. For example, if a method throws an
exception annotated with @ResponseStatus there should be no need to
document that possible response anywhere else, or if a method
parameter is already described in Javadoc then that is sufficient for
the parameter to be documented in the REST API documentation.

Spring Boot support is provided. This support auto-configures
endpoints that will serve raw JSON describing the API and Swagger's
HTML-based documentation.